chore_: bridge moved to router
