+++ services/wallet/router/filter_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
 	"github.com/status-im/status-go/params"
-	"github.com/status-im/status-go/services/wallet/bridge"
+	"github.com/status-im/status-go/services/wallet/router/bridge"
 
 	"github.com/stretchr/testify/assert"
 )
