+++ services/wallet/transfer/helpers.go
@@ -14,8 +14,8 @@ import (
 	"github.com/status-im/status-go/account"
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/services/wallet/bridge"
 	wallet_common "github.com/status-im/status-go/services/wallet/common"
+	"github.com/status-im/status-go/services/wallet/router/bridge"
 )
 
 func rowsToMultiTransactions(rows *sql.Rows) ([]*MultiTransaction, error) {
