+++ services/wallet/transfer/transaction_manager_multitransaction_test.go
@@ -16,9 +16,9 @@ import (
 	"github.com/status-im/status-go/account"
 	"github.com/status-im/status-go/eth-node/types"
 	"github.com/status-im/status-go/rpc"
-	"github.com/status-im/status-go/services/wallet/bridge"
-	"github.com/status-im/status-go/services/wallet/bridge/mock_bridge"
 	wallet_common "github.com/status-im/status-go/services/wallet/common"
+	"github.com/status-im/status-go/services/wallet/router/bridge"
+	"github.com/status-im/status-go/services/wallet/router/bridge/mock_bridge"
 	"github.com/status-im/status-go/services/wallet/walletevent"
 	"github.com/status-im/status-go/t/helpers"
 	"github.com/status-im/status-go/transactions"
